Fix more todos
