+++ src/certificate.rs
@@ -1,0 +1,1 @@
+pub use hotshot_types::certificate::QuorumCertificate;
