+++ src/traits/storage/memory_storage.rs
@@ -121,9 +121,9 @@ mod test {
     use crate::traits::election::static_committee::StaticVoteToken;
 
     use super::*;
+    use hotshot_types::certificate::QuorumCertificate;
     use hotshot_types::constants::genesis_proposer_id;
     use hotshot_types::data::fake_commitment;
-    use hotshot_types::data::QuorumCertificate;
     use hotshot_types::data::{ValidatingLeaf, ViewNumber};
     #[allow(clippy::wildcard_imports)]
     use hotshot_types::traits::block_contents::dummy::*;
